requirements.txt
streamlit
pandas
snowflake-connector-python
google-generativeai
matplotlib
seaborn
